# backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: StreamPulse API
  version: 0.1.0
  description: REST API for the StreamPulse backend.
servers:
  - url: /
...
paths:
...
  /api/tracks:
    get:
      tags:
        - Track
      operationId: listUserTracks
      summary: List the authenticated user's track library.
      description: >-
        Returns the tracks owned by the authenticated user, most recent first.
        It is the source of the "add a track" picker (US-05-03).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The user's tracks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Track
      operationId: uploadTrack
      summary: Upload an audio file to the personal library.
      description: >-
        Uploads an audio file (MP3/AAC/OGG, ≤ 50 MB) to the authenticated user's
        library (US-05-01). The real MIME type is sniffed server-side (a
        non-audio file, e.g. a PDF renamed .mp3, is rejected with 415); the file
        is stored outside any served directory and referenced in the database
        with its metadata. Duration is optional and supplied by the client.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - title
              properties:
                file:
                  type: string
                  format: binary
                  description: The audio file (MP3, AAC or OGG), ≤ 50 MB.
                title:
                  type: string
                  description: Track title.
                artist:
                  type: string
                  description: Optional artist name.
                duration_s:
                  type: integer
                  description: Optional track duration in seconds (positive).
      responses:
        "201":
          description: The created track.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The user's cumulated storage quota would be exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                quotaExceeded:
                  value:
                    error:
                      code: storage_quota_exceeded
                      message: quota de stockage atteint
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          description: The uploaded file exceeds the 50 MB limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tooLarge:
                  value:
                    error:
                      code: file_too_large
                      message: le fichier dépasse la taille maximale de 50 Mo
        "415":
          description: The file is not an accepted audio type (MP3/AAC/OGG).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notAudio:
                  value:
                    error:
                      code: unsupported_media_type
                      message: seuls les fichiers audio MP3, AAC ou OGG sont acceptés
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tracks/{id}/stream:
    get:
      tags:
        - Track
      operationId: streamTrack
      summary: Stream the audio file of one of the user's tracks.
      description: >-
        Serves the raw audio bytes of a track owned by the authenticated user
        (US-05-04), so the mobile player can play a playlist queue. Range
        requests are honoured (`Accept-Ranges: bytes`), which lets the player
        resume or seek within a track without downloading it again. A track that
        belongs to somebody else is reported as 404, exactly like an unknown one:
        the status code never reveals that it exists.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Track identifier.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The audio file.
          headers:
            Accept-Ranges:
              description: Always `bytes` — the player may request slices.
              schema:
                type: string
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/aac:
              schema:
                type: string
                format: binary
            audio/ogg:
              schema:
                type: string
                format: binary
        "206":
          description: The requested byte range of the audio file.
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/aac:
              schema:
                type: string
                format: binary
            audio/ogg:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request payload or invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidJson:
              value:
                error:
                  code: invalid_json
                  message: invalid json body
            invalidArgument:
              value:
                error:
                  code: invalid_argument
                  message: invalid email
    Unauthorized:
      description: Authentication failed or bearer token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                error:
                  code: unauthorized
                  message: invalid credentials
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            duplicate:
              value:
                error:
                  code: conflict
                  message: email or username already taken
...
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: resource not found
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error:
                  code: internal
                  message: internal server error
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: invalid_argument
        message:
          type: string
          example: invalid email
...
    TrackResponse:
      type: object
      required:
        - id
        - title
        - artist
        - duration_s
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Midnight Drive
        artist:
          type: string
          nullable: true
          example: Neon Lights
        duration_s:
          type: integer
          nullable: true
          description: Track duration in seconds.
          example: 214
...